refactor: Entity 수정
